Move comments into web-comments
